renamed custom_logging to logger
